main updated

+++ Search.h
@@ -1,3 +1,3 @@
 #include "Position.h"
 int SearchMain(PositionStruct& pos, int gotime);
-extern int hsh[15][100];
+extern long long hsh[15][100];
